--- KeySwitchManager/Sources/Runtime/Infrastructures/Storage.Xml/Cubase/Models/ExtraDataKeys.cs
@@ -1,12 +1,0 @@
-using KeySwitchManager.Domain.KeySwitches.Models.Values;
-
-namespace KeySwitchManager.Infrastructures.Storage.Xml.Cubase.Models
-{
-    public static class ExtraDataKeys
-    {
-        public static readonly ExtraDataKey SlotName = new ExtraDataKey( "Cubase.Slot" );
-        public static readonly ExtraDataKey ColorIndex = new ExtraDataKey( "Cubase.Color" );
-        public static readonly ExtraDataKey GroupIndex = new ExtraDataKey( "Cubase.Group" );
-        public static readonly ExtraDataKey ArticulationType = new ExtraDataKey( "Cubase.ArticulationType" );
-    }
-}
