--- KeySwitchManager/Sources/Runtime/Infrastructures/Storage.Xml/Cubase/MultipleCubaseWriter.cs
@@ -1,40 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-using KeySwitchManager.Commons.Data;
-using KeySwitchManager.Domain.KeySwitches.Models;
-using KeySwitchManager.Infrastructures.Storage.KeySwitches.Helper;
-
-namespace KeySwitchManager.Infrastructures.Storage.Xml.Cubase
-{
-    public class MultipleCubaseWriter : IKeySwitchWriter
-    {
-        private const string Suffix = ".expressionmap";
-
-        public DirectoryPath OutputDirectory { get; }
-        private Encoding FileEncoding { get; }
-        public bool LeaveOpen => false;
-
-        public MultipleCubaseWriter( DirectoryPath outputDirectory ) : this( outputDirectory, Encoding.UTF8 ) {}
-
-        public MultipleCubaseWriter( DirectoryPath outputDirectory, Encoding filEncoding )
-        {
-            OutputDirectory = outputDirectory;
-            FileEncoding    = filEncoding;
-        }
-
-        public void Dispose() {}
-
-        public void Write( IReadOnlyCollection<KeySwitch> keySwitches, IObserver<string>? loggingSubject = null )
-        {
-            MultipleWritingHelper.Write(
-                keySwitches,
-                OutputDirectory,
-                Suffix,
-                loggingSubject,
-                stream => new CubaseWriter( stream, FileEncoding, true )
-            );
-        }
-    }
-}
