--- KeySwitchManager/Sources/Runtime/Infrastructures/Storage.Xml/StudioOne/Translators/StudioOneExportTranslator.cs
@@ -1,18 +1,0 @@
-using KeySwitchManager.Commons.Data;
-using KeySwitchManager.Domain.KeySwitches.Models;
-using KeySwitchManager.Infrastructures.Storage.Xml.StudioOne.Translators.Helpers;
-
-using RkHelper.Text.Xml;
-
-namespace KeySwitchManager.Infrastructures.Storage.Xml.StudioOne.Translators
-{
-    public class StudioOneExportTranslator : IDataTranslator<KeySwitch, IText>
-    {
-        public IText Translate( KeySwitch source )
-        {
-            return new PlainText(
-                XmlHelper.ToXmlString( KeySwitchToStudioOneModelHelper.Translate( source ) )
-            );
-        }
-    }
-}
